added check user role feature

+++ server/src/controllers/ActivityCard/deleteActivityCard.js
@@ -2,9 +2,16 @@ import express from 'express'
 import mongoose from 'mongoose'
 
 import { ActivityCard } from '../../models/index.js'
+import clerk from '../../middleware/clerk.js'
 
 const deleteActivityCard = async (req, res) => {
   const { id } = req.params
+  const { userId } = req.auth
+  const user = await clerk.users.getUser(userId)
+
+  if(!user || user.publicMetadata.role !== 'copywriter') {
+    return res.status(403).json({ message: 'Unauthorized' })
+  }
 
   try {
     await ActivityCard.findByIdAndRemove(id)
